old/my-webapp/requirements.txt
Flask
python-dotenv
pyotp
logzero
SmartApi
requests
Jinja2